Refactors and expands VisualizationConverter
Moves the `VisualizationConverter` to the Core utilities and expands its functionality to handle various visualization types. This change enables the conversion of a wider range of visualizations to GridVisualizations, including financial charts, gauges, maps, and KPI visualizations. It also updates the GridVisualization to make use of the converter and provide an example.
