+++ packages/dom/src/Core/Utilities/index.ts
@@ -1,1 +1,2 @@
 export { RvDashboardLoader } from "./RvDashboardLoader";
+export { VisualizationConverter } from "./VisualizationConverter";
