+++ packages/dom/src/Visualizations/GridVisualization.ts
@@ -7,7 +7,7 @@ import { Visualization } from "./Visualization";
 import { ColumnUtilities } from "./Utilities/ColumnUtilities";
 import { GridVisualizationDataSpec } from "./VisualizationSpecs/GridVisualizationDataSpec";
 import { IVisualization } from "./Interfaces/IVisualization";
-import { VisualizationConverter, GridConversionOptions } from "./Converters/VisualizationConverter";
+import { VisualizationConverter, GridConversionOptions } from "../Core/Utilities/VisualizationConverter";
 
 export class GridVisualization extends Visualization<GridVisualizationSettings> {
 
@@ -17,6 +17,18 @@ export class GridVisualization extends Visualization<GridVisualizationSettings> 
         this.chartType = ChartType.Grid;
     }
 
+    /**
+     * Converts a visualization to a GridVisualization.
+     * @param sourceViz The source visualization to convert.
+     * @param options Optional conversion options.
+     * @returns A new GridVisualization instance with the converted data, or null if the visualization cannot be converted (e.g., TextBox, Custom, Image, or Grid types).
+     * @example
+     * // Convert with only visualization fields
+     * const grid = GridVisualization.from(barChart);
+     * 
+     * // Convert with all data source fields
+     * const grid = GridVisualization.from(barChart, { includeAllFields: true });
+     */
     static from(sourceViz: IVisualization, options?: GridConversionOptions): GridVisualization | null {
         return VisualizationConverter.toGrid(sourceViz, options);
     }
